requirements.txt
crewai
crewai_tools
langchain-huggingface
python-dotenv
youtube-transcript-api
